refactored gates copy to align with redeem code cta
